# bloxz-backend-master/api-spec.yml
openapi: 3.0.1
info:
  title: bloxz Event API Spefication
  version: 0.0.1
servers:
  - url: https://localhost:5001/api

tags:
  - name: event

paths:
  /events:
    get:
      summary: Get all events
      tags:
        - event
      requestBody:
        description: OK
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Event"
      responses:
        "200":
          $ref: "#/components/schemas/OK"
        "500":
          $ref: "#/components/schemas/InternalServerError"

    post:
      summary: Create event
      tags:
        - event
      requestBody:
        description: The event object to be created.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Event"
        required: true
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/schemas/InternalServerError"

  /events/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get event
      tags:
        - event
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/schemas/InternalServerError"

    put:
      summary: Update event
      tags:
        - event
      requestBody:
        description: The updated event object.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Event"
        required: true
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/schemas/InternalServerError"

    delete:
      summary: Delete event
      tags:
        - event
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/schemas/InternalServerError"

components:
  schemas:
    Event:
      type: object
      required:
        - title
        - description
        - location
        - start
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        location:
          type: string
        start:
          type: string
          format: date-time
  responses:
    Created:
      description: Created
      headers:
        Location:
          description: The location of the created object.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Event"
    NoContent:
      description: No Content
    BadRequest:
      description: Bad Request. Invalid event supplied.
    NotFound:
      description: Not Found
    InternalServerError:
      description: Internal Server Error